requirements.txt
discord.py
requests
python-dotenv
supabase
pytz
aiohttp
flask
rapidfuzz
